refactor!: reorganize package exports under tailwind/ panda/ stylex/ folders
BREAKING CHANGE: every public subpath has moved.

Old → new:
  @klinking/squircle/tw-plugin        → @klinking/squircle/tailwind
  @klinking/squircle/tw-merge-cfg     → @klinking/squircle/tailwind
                                        (now exports `squircleMergeConfig`
                                        as a named export from the same file)
  @klinking/squircle/tw-utils.css     → @klinking/squircle/tailwind/utils.css
  @klinking/squircle/squircle-radius.css → @klinking/squircle/tailwind/radius.css
  @klinking/squircle/panda-preset     → @klinking/squircle/panda
  @klinking/squircle/stylex           → @klinking/squircle/stylex (unchanged)

The Tailwind plugin and tailwind-merge config now ship from a single file
(`dist/tailwind/index.mjs`) — the plugin is the default export, the merge
config is a named export. Source files are renamed to match
(src/tailwind.ts, src/panda.ts, src/stylex.ts).

Updated:
- package.json exports
- vite-plus pack entries (folder/index pattern)
- generate-squircle-css.ts (writes into dist/tailwind/)
- test-utils.ts, squircle-radius.test.ts (new dist paths)
- website + examples imports
- README + sync-readme.sh (re-synced appendix to the new layout, added
  Path D for the StyleX integration)

All 183 package tests pass and the website builds + dev-renders against
the new dist layout.
