+++ website/examples/panda/panda.config.ts
@@ -1,5 +1,5 @@
 import { defineConfig } from "@pandacss/dev";
-import squirclePreset from "@klinking/squircle/panda-preset";
+import squirclePreset from "@klinking/squircle/panda";
 
 export default defineConfig({
   prefix: "pd",
